+++ tests/BlueprintTest.php
@@ -4,8 +4,6 @@ namespace Rougin\Blueprint;
 
 use Symfony\Component\Console\Tester\CommandTester;
 
-use Rougin\Blueprint\Console;
-
 /**
  * Blueprint Test
  *
@@ -15,27 +13,19 @@ use Rougin\Blueprint\Console;
 class BlueprintTest extends \PHPUnit_Framework_TestCase
 {
     /**
-     * Sets up the filename to be used.
-     *
-     * @return void
-     */
-    public function setUp()
-    {
-        if (! defined('BLUEPRINT_FILENAME')) {
-            define('BLUEPRINT_FILENAME', 'blueprint.yml');
-        }
-    }
-
-    /**
      * Tests \Rougin\Blueprint\Blueprint::run.
      *
      * @return void
      */
     public function testRun()
     {
-        $console = Console::boot(BLUEPRINT_FILENAME, new \Auryn\Injector, getcwd())->run(true);
+        $console = Console::boot(null, null, getcwd());
 
-        $this->assertInstanceOf('Symfony\Component\Console\Application', $console);
+        $expected = 'Symfony\Component\Console\Application';
+
+        $result = $console->run(true);
+
+        $this->assertInstanceOf($expected, $result);
     }
 
     /**
@@ -45,16 +35,17 @@ class BlueprintTest extends \PHPUnit_Framework_TestCase
      */
     public function testInitializeCommand()
     {
-        $blueprint = Console::boot(BLUEPRINT_FILENAME);
-        $className = 'Rougin\Blueprint\Commands\InitializeCommand';
-        $instance  = $blueprint->injector->make($className);
+        $blueprint = Console::boot(null, null, __DIR__);
 
-        $command = new CommandTester($instance);
-        $command->execute([]);
+        $init = $blueprint->run(true)->find('init');
 
-        $this->assertFileExists(BLUEPRINT_FILENAME);
+        $command = new CommandTester($init);
 
-        unlink(getcwd() . '/' . BLUEPRINT_FILENAME);
+        $command->execute(array());
+
+        $this->assertFileExists(__DIR__ . '/blueprint.yml');
+
+        unlink(__DIR__ . '/blueprint.yml');
     }
 
     /**
@@ -64,39 +55,37 @@ class BlueprintTest extends \PHPUnit_Framework_TestCase
      */
     public function testSetTemplatePath()
     {
-        $separator = DIRECTORY_SEPARATOR;
-        $templates = str_replace('tests', 'src', __DIR__) . $separator . 'Templates';
+        $root = str_replace('tests', 'src', __DIR__);
 
-        $blueprint = Console::boot(BLUEPRINT_FILENAME);
+        $expected = $root . '/Templates';
 
-        $blueprint->setTemplatePath($templates);
+        $blueprint = Console::boot('blueprint.yml');
 
-        $this->assertEquals($templates, $blueprint->getTemplatePath());
+        $blueprint->setTemplatePath($expected);
+
+        $result = $blueprint->getTemplatePath();
+
+        $this->assertEquals($expected, $result);
     }
 
     /**
-     * Tests Rougin\Blueprint\Commands\GreetCommand.
+     * Tests Rougin\Blueprint\Fixture\GreetCommand.
      *
      * @return void
      */
-    public function testGreetCommand()
-    {
-        $blueprint = Console::boot(__DIR__ . '/blueprint.yml');
-
-        $className = 'Rougin\Blueprint\Commands\InitializeCommand';
-        $instance  = $blueprint->injector->make($className);
-
-        $command = new CommandTester($instance);
-        $command->execute([]);
-
-        $className = 'Rougin\Blueprint\Commands\GreetCommand';
-        $instance  = $blueprint->injector->make($className);
-
-        $command = new CommandTester($instance);
-        $command->execute([ 'name' => 'Rougin', '--yell' => true ]);
-
-        $this->assertRegExp('/HELLO ROUGIN!/', $command->getDisplay());
-
-        unlink(BLUEPRINT_FILENAME);
-    }
+    // public function testGreetCommand()
+    // {
+    //     $blueprint = Console::boot(__DIR__ . '/blueprint.yml');
+    //     $blueprint = Console::boot(__DIR__ . '/blueprint.yml');
+    //     $className = 'Rougin\Blueprint\Commands\InitializeCommand';
+    //     $instance  = $blueprint->injector->make($className);
+    //     $command = new CommandTester($instance);
+    //     $command->execute([]);
+    //     $className = 'Rougin\Blueprint\Commands\GreetCommand';
+    //     $instance  = $blueprint->injector->make($className);
+    //     $command = new CommandTester($instance);
+    //     $command->execute([ 'name' => 'Rougin', '--yell' => true ]);
+    //     $this->assertRegExp('/HELLO ROUGIN!/', $command->getDisplay());
+    //     unlink(BLUEPRINT_FILENAME);
+    // }
 }